+++ src/TextShow/TH/Internal.hs
@@ -92,7 +92,7 @@ import           GHC.Exts ( Char(..), Double(..), Float(..), Int(..), Word(..)
                           )
 import           GHC.Show (appPrec, appPrec1)
 
-import           Language.Haskell.TH.Datatype
+import           Language.Haskell.TH.Datatype as Datatype
 import           Language.Haskell.TH.Lib
 import           Language.Haskell.TH.Ppr hiding (appPrec)
 import           Language.Haskell.TH.Syntax
@@ -874,14 +874,17 @@ buildTypeInstance tsClass tyConName dataCxt varTysOrig variant = do
           map (substNamesWithKindStar (List.union droppedKindVarNames kvNames'))
             $ take remainingLength varTysOrig
 
-        isDataFamily :: Bool
-        isDataFamily = case variant of
-                         Datatype        -> False
-                         Newtype         -> False
-                         DataInstance    -> True
-                         NewtypeInstance -> True
+    isDataFamily <-
+      case variant of
+        Datatype        -> return False
+        Newtype         -> return False
+        DataInstance    -> return True
+        NewtypeInstance -> return True
+#if MIN_VERSION_th_abstraction(0,5,0)
+        Datatype.TypeData -> typeDataError tyConName
+#endif
 
-        remainingTysOrigSubst' :: [Type]
+    let remainingTysOrigSubst' :: [Type]
         -- See Note [Kind signatures in derived instances] for an explanation
         -- of the isDataFamily check.
         remainingTysOrigSubst' =
@@ -1061,6 +1064,17 @@ outOfPlaceTyVarError tsClass conName = Monad.fail
   where
     n :: Int
     n = fromEnum tsClass
+
+#if MIN_VERSION_th_abstraction(0,5,0)
+-- | We cannot implement class methods at the term level for @type data@
+-- declarations, which only exist at the type level.
+typeDataError :: Name -> Q a
+typeDataError dataName = fail
+  . showString "Cannot derive instance for ‘"
+  . showString (nameBase dataName)
+  . showString "‘, which is a ‘type data‘ declaration"
+  $ ""
+#endif
 
 -------------------------------------------------------------------------------
 -- Expanding type synonyms